requirements.txt
mkdocs==1.6.1
mkdocs-material~=9.7.6
mkdocs-material-extensions==1.3.1
mkdocs-roamlinks-plugin==0.3.2
mkdocs-git-revision-date-localized-plugin
mkdocs-macros-plugin
pyyaml==6.0.3
tenacity==9.0.0
